Fixed a couple of small bugs, got this running on the new data
And, because I put the functionality of get_mids_for_sfe in the
TrainingDataProcessor, that file is no longer needed.  The behavior in
TrainingDataProcessor isn't identical to what was in get_mids_for_sfe, but I
think the new version is the correct way to do it.  And because we can do
feature computation on the fly now during training, having all of the mid
features precomputed isn't as big a deal - this is just for selecting word
features using the PMI pipeline, and I think the current behavior is correct
for that case.